Refactoring for 0.5.0
Structure
- all configs switched to JSON
- renamed 'templates' to 'html'
- split media components into separate files

ESLINT
- 'eol-last' switched to 'always' from 'never'
- removed sort imports
