+++ app-src/configs/config.json
@@ -1,0 +1,19 @@
+{
+	"appName": "loomSE",
+	"appRoot": "loomSE",
+	"version": "0.5.0",
+	"scripts": {
+		"mobile": "assets/scripts/script-mobile.json",
+		"desktop": "assets/scripts/script-desktop.json"
+	},
+	"mobile": {
+		"minimumResolution": 480
+	},
+	"showLoadingPage": true,
+	"developer": {
+		"mute": true,
+		"disableCheckScript": false,
+		"disableScrubScreen": false,
+		"verbose": []
+	}
+}
