+++ app-src/html/loading.html
@@ -1,0 +1,1 @@
+<h1>Loading...</h1>
