+++ app-src/html/unableToPlay.html
@@ -1,0 +1,1 @@
+<h1>Unable to play on your device</h1>
