+++ app-src/model/scriptHandler.js
