--- app-src/templates/loading.html
@@ -1,1 +1,0 @@
-<h1>Loading...</h1>
