--- app-src/templates/unableToPlay.html
@@ -1,1 +1,0 @@
-<h1>Unable to play on your device</h1>
