+++ app-src/tools/common.js
