+++ app-src/view/components/media/index.js
@@ -1,0 +1,1 @@
+export { default as media } from './media';
