+++ app-src/view/components/media_gui/html/media_gui.html
