+++ app-src/view/components/media_gui/index.js
@@ -1,0 +1,1 @@
+export { default as mediaGui } from './media_gui';
