+++ index.html
@@ -6,7 +6,7 @@
 		<title>Loom Test Card</title>
 		<link type="text/css" href="styles/master.css" rel="stylesheet" />
 
-		<script type="text/javascript" src="app-build/loomSE-0.4.2.js"></script>
+		<script type="text/javascript" src="app-build/loomSE-0.5.0.js"></script>
 	</head>
 
 	<body>
