+++ package.json
@@ -1,7 +1,7 @@
 {
   "private": true,
   "name": "loomse",
-  "version": "0.4.2",
+  "version": "0.5.0",
   "description": "Interactive storytelling for the modern web",
   "scripts": {
     "watch": "./node_modules/.bin/rollup -c --watch;",
@@ -22,6 +22,7 @@
     "rollup-plugin-babel": "^2.7.1",
     "rollup-plugin-eslint": "^3.0.0",
     "rollup-plugin-html": "^0.2.1",
+    "rollup-plugin-json": "^2.3.0",
     "rollup-plugin-replace": "^1.1.1",
     "rollup-plugin-uglify": "^1.0.1"
   },
